+++ src/svf/downloader.ts
@@ -2,9 +2,8 @@ import * as path from 'path';
 import * as fse from 'fs-extra';
 import axios from 'axios';
 import { SvfReader } from '..';
-import { SdkManager, SdkManagerBuilder } from '@aps_sdk/autodesk-sdkmanager';
 import { IAuthenticationProvider } from '../common/authentication-provider';
-import { ManifestDerivativesChildren, ModelDerivativeClient } from '@aps_sdk/model-derivative';
+import { ManifestResources, ModelDerivativeClient } from '@aps_sdk/model-derivative';
 import { Scopes } from '@aps_sdk/authentication';
 
 export interface IDownloadOptions {
@@ -26,12 +25,10 @@ interface IDownloadContext {
 }
 
 export class Downloader {
-    protected sdkManager: SdkManager;
     protected modelDerivativeClient: ModelDerivativeClient;
 
     constructor(protected authenticationProvider: IAuthenticationProvider, host?: string, region?: string) {
-        this.sdkManager = SdkManagerBuilder.create().build();
-        this.modelDerivativeClient = new ModelDerivativeClient(this.sdkManager);
+        this.modelDerivativeClient = new ModelDerivativeClient();
     }
 
     download(urn: string, options?: IDownloadOptions): IDownloadTask {
@@ -49,7 +46,7 @@ export class Downloader {
 
     private async _downloadDerivative(urn: string, derivativeUrn: string) {
         const accessToken = await this.authenticationProvider.getToken([Scopes.ViewablesRead]);
-        const downloadInfo = await this.modelDerivativeClient.getDerivativeUrl(accessToken, derivativeUrn, urn);
+        const downloadInfo = await this.modelDerivativeClient.getDerivativeUrl(derivativeUrn, urn, { accessToken });
         const response = await axios.get(downloadInfo.url as string, { responseType: 'arraybuffer', decompress: false });
         return response.data;
     }
@@ -57,11 +54,11 @@ export class Downloader {
     private async _download(urn: string, context: IDownloadContext): Promise<void> {
         context.log(`Downloading derivative ${urn}`);
         const accessToken = await this.authenticationProvider.getToken([Scopes.ViewablesRead]);
-        const manifest = await this.modelDerivativeClient.getManifest(accessToken, urn);
+        const manifest = await this.modelDerivativeClient.getManifest(urn, { accessToken });
         const urnDir = path.join(context.outputDir || '.', urn);
 
-        const derivatives: ManifestDerivativesChildren[] = [];
-        function collectDerivatives(derivative: ManifestDerivativesChildren) {
+        const derivatives: ManifestResources[] = [];
+        function collectDerivatives(derivative: ManifestResources) {
             if (derivative.type === 'resource' && derivative.role === 'graphics' && (derivative as any).mime === 'application/autodesk-svf') {
                 derivatives.push(derivative);
             }