+++ samples/shared.js
@@ -6,10 +6,18 @@ const { ModelDerivativeClient } = require('@aps_sdk/model-derivative');
 async function downloadDerivative(urn, derivativeUrn, clientId, clientSecret) {
     const authenticationClient = new AuthenticationClient(SdkManagerBuilder.create().build());
     const modelDerivativeClient = new ModelDerivativeClient();
-    const credentials = await authenticationClient.getTwoLeggedToken(clientId, clientSecret, [Scopes.ViewablesRead]);
-    const downloadInfo = await modelDerivativeClient.getDerivativeUrl(derivativeUrn, urn, { accessToken: credentials.access_token });
-    const response = await axios.get(downloadInfo.url, { responseType: 'arraybuffer', decompress: false });
-    return response.data;
+    try {
+        const credentials = await authenticationClient.getTwoLeggedToken(clientId, clientSecret, [Scopes.ViewablesRead]);
+        const downloadInfo = await modelDerivativeClient.getDerivativeUrl(derivativeUrn, urn, { accessToken: credentials.access_token });
+        const response = await axios.get(downloadInfo.url, { responseType: 'arraybuffer', decompress: false });
+        return response.data;
+    } catch (error) {
+        if (axios.isAxiosError(error)) {
+            throw new Error(`Could not download derivative ${derivativeUrn}: ${error.message}`);
+        } else {
+            throw error;
+        }
+    }
 }
 
 async function getSvfDerivatives(urn, clientId, clientSecret) {