+++ src/svf/reader.ts
@@ -237,10 +237,18 @@ export class Reader {
         }
 
         async function downloadDerivative(urn: string, derivativeUrn: string) {
-            const accessToken = await authenticationProvider.getToken([Scopes.ViewablesRead]);
-            const downloadInfo = await modelDerivativeClient.getDerivativeUrl(derivativeUrn, urn, { accessToken });
-            const response = await axios.get(downloadInfo.url as string, { responseType: 'arraybuffer', decompress: false });
-            return response.data;
+            try {
+                const accessToken = await authenticationProvider.getToken([Scopes.ViewablesRead]);
+                const downloadInfo = await modelDerivativeClient.getDerivativeUrl(derivativeUrn, urn, { accessToken });
+                const response = await axios.get(downloadInfo.url as string, { responseType: 'arraybuffer', decompress: false });
+                return response.data;
+            } catch (error) {
+                if (axios.isAxiosError(error)) {
+                    throw new Error(`Could not download derivative ${derivativeUrn}: ${error.message}`);
+                } else {
+                    throw error;
+                }
+            }
         }
 
         const svfUrn = (foundDerivative as any).urn;