+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Modified
+  - Upgraded to newer version of APS SDK
+
 ## [5.0.3] - 2024-04-09
 
 - Modified