+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.4] - 2024-09-10
+
+- Added
+  - Cleaning up error logs for Axios-related errors
+- Modified
+  - Upgraded to newer version of APS SDK
+
 ## [5.0.3] - 2024-04-09
 
 - Modified